+++ src/tests/FeatureTests/BigInt64Array.js
@@ -1,0 +1,110 @@
+const test = require('tape');
+const testHelpers = require('/tests/testHelpers.js');
+const jsonComplete = require('/main.js');
+
+const encode = jsonComplete.encode;
+const decode = jsonComplete.decode;
+
+if (typeof BigInt64Array === 'function') {
+    test('BigInt64Array: Normal', (t) => {
+        t.plan(3);
+
+        const a = new BigInt64Array(2);
+        a[0] = BigInt(-1);
+        a[1] = BigInt(2);
+
+        const decoded = decode(encode([a]))[0];
+
+        t.equal(testHelpers.systemName(decoded), '[object BigInt64Array]');
+        t.equal(decoded[0], BigInt(-1));
+        t.equal(decoded[1], BigInt(2));
+    });
+
+    test('BigInt64Array: Empty Cells', (t) => {
+        t.plan(2);
+
+        const a = new BigInt64Array(2);
+        a[0] = BigInt(1);
+
+        const decoded = decode(encode([a]))[0];
+
+        t.equal(decoded[0], BigInt(1));
+        t.equal(decoded[1], BigInt(0));
+    });
+
+    test('BigInt64Array: Empty', (t) => {
+        t.plan(2);
+
+        const a = new BigInt64Array(0);
+
+        const decoded = decode(encode([a]))[0];
+
+        t.equal(testHelpers.systemName(decoded), '[object BigInt64Array]');
+        t.equal(decoded.length, 0);
+    });
+
+    test('BigInt64Array: Root Value', (t) => {
+        t.plan(1);
+        t.deepEqual(decode(encode(new BigInt64Array([BigInt(1)]))), new BigInt64Array([BigInt(1)]));
+    });
+
+    test('BigInt64Array: Arbitrary Attached Data', (t) => {
+        t.plan(2);
+
+        const a = new BigInt64Array(2);
+        a.x = 2;
+        a[Symbol.for('BigInt64Array')] = 'test';
+
+        const decoded = decode(encode([a], {
+            encodeSymbolKeys: true,
+        }))[0];
+
+        t.equal(decoded.x, 2);
+        t.equal(decoded[Symbol.for('BigInt64Array')], 'test');
+    });
+
+    test('BigInt64Array: Self-Containment', (t) => {
+        t.plan(1);
+
+        const a = new BigInt64Array(2);
+        a.me = a;
+
+        const decoded = decode(encode([a]))[0];
+
+        t.equal(decoded.me, decoded);
+    });
+
+    test('BigInt64Array: Referential Integrity', (t) => {
+        t.plan(2);
+
+        const source = new BigInt64Array(1);
+
+        const decoded = decode(encode({
+            x: source,
+            y: source,
+        }));
+
+        t.equal(decoded.x, decoded.y);
+        t.notEqual(decoded.x, source);
+    });
+
+    test('BigInt64Array: Encoding Expected', (t) => {
+        t.plan(1);
+
+        const source = new BigInt64Array(1);
+        source[0] = BigInt(1);
+        source.a = false;
+
+        t.deepEqual(testHelpers.simplifyEncoded(encode(source)), {
+            BI: '_0 S0 F0',
+            _: '1',
+            S: [
+                'a',
+            ],
+            r: 'BI0',
+        });
+    });
+}
+else {
+    console.warn('Tests for BigInt64Array type skipped because it is not supported in the current environment.'); // eslint-disable-line no-console
+}