+++ src/tests/FeatureTests/DecodingSymbolKeyWhenSymbolsNotSupported.js
@@ -4,7 +4,13 @@ const jsonComplete = require('/main.js');
 
 const decode = jsonComplete.decode;
 
-const encodedSymbolKeyObject = '[["Ob",[[["St0","Sy0"],["St1","St2"]]]],["St",["k","2","1"]],["Sy",["Rs"]],["r","Ob0"],["v","1.0.0"]]';
+// encode({
+//     k: 2,
+//     [Symbol.for('s')]: 1,
+// }, {
+//     encodeSymbolKeys: true,
+// });
+const encodedSymbolKeyObject = '["O0","2.0.0",["O","S0P0 N0N1"],["S",["k"]],["P",["rs"]],["N","2,1"]]';
 
 test('Decoding Symbol Key when Symbols are Not Supported: Defaults to throw', (t) => {
     t.plan(2);
@@ -22,7 +28,7 @@ test('Decoding Symbol Key when Symbols are Not Supported: Defaults to throw', (t
         console.log(decoded); // eslint-disable-line no-console
         t.ok(false);
     } catch (e) {
-        t.equal(e.message, 'Cannot decode recognized pointer type "Sy".');
+        t.equal(e.message, 'Cannot decode recognized pointer type "P".');
     }
 
     globalThis.Symbol = oldSymbol;
@@ -43,7 +49,7 @@ test('Decoding Symbol Key when Symbols are Not Supported: Skips Symbol Keys in c
         const decoded = decode(encodedSymbolKeyObject, {
             compat: true,
         });
-        t.deepEqual(decoded, {k: '2'});
+        t.deepEqual(decoded, {k: 2});
     } catch (e) {
         t.ok(false);
         throw e;