+++ src/tests/FeatureTests/Error.js
@@ -167,35 +167,22 @@ test('Error: Encoding Expected', (t) => {
     const encoded = testHelpers.simplifyEncoded(encode(value));
 
     // Simplify stack for check
-    encoded.St[2] = 'stack!!!';
+    encoded.S[2] = 'stack!!!';
 
     // Remove extra properties a given browser might have added (namely, Safari)
-    delete encoded.Nu;
-    encoded.St = encoded.St.slice(0, 3);
-    encoded.Er[0][1] = encoded.Er[0][1].slice(0, 1);
-    encoded.Er[0][2] = encoded.Er[0][2].slice(0, 1);
+    delete encoded.N;
+    encoded.S = encoded.S.slice(0, 3);
+    encoded.E = encoded.E.split(' ').map((parts, index) => {
+        return index === 0 ? parts : parts.slice(0, 2);
+    }).join(' ');
 
     t.deepEqual(encoded, {
-        Er: [
-            [
-                [
-                    'St0',
-                    'St1',
-                    'St2',
-                ],
-                [
-                    'St1',
-                ],
-                [
-                    'fa',
-                ],
-            ],
-        ],
-        St: [
+        E: 'S0S1S2 S1 F0',
+        S: [
             'Error',
             'a',
             'stack!!!',
         ],
-        r: 'Er0',
+        r: 'E0',
     });
 });