+++ src/tests/FeatureTests/Float32Array.js
@@ -96,26 +96,12 @@ if (typeof Float32Array === 'function') {
         source.a = false;
 
         t.deepEqual(testHelpers.simplifyEncoded(encode(source)), {
-            F3: [
-                [
-                    [
-                        'Nu0',
-                    ],
-                    [
-                        'St0',
-                    ],
-                    [
-                        'fa',
-                    ],
-                ],
+            FT: 'N0 S0 F0',
+            N: '1',
+            S: [
+                'a'
             ],
-            Nu: [
-                '1',
-            ],
-            St: [
-                'a',
-            ],
-            r: 'F30',
+            r: 'FT0',
         });
     });
 }