+++ src/tests/FeatureTests/ObjectWrappedNumber.js
@@ -139,26 +139,12 @@ test('Object-Wrapped Number: Encoding Expected Normal', (t) => {
     source.a = false;
 
     t.deepEqual(testHelpers.simplifyEncoded(encode(source)), {
-        NU: [
-            [
-                [
-                    'Nu0',
-                ],
-                [
-                    'St0',
-                ],
-                [
-                    'fa',
-                ],
-            ],
-        ],
-        Nu: [
-            '1',
-        ],
-        St: [
+        H: 'N0 S0 F0',
+        N: '1',
+        S: [
             'a',
         ],
-        r: 'NU0',
+        r: 'H0',
     });
 });
 
@@ -168,14 +154,8 @@ test('Object-Wrapped Number: Encoding Expected Infinity', (t) => {
     const source = new Number(Infinity);
 
     t.deepEqual(testHelpers.simplifyEncoded(encode(source)), {
-        NU: [
-            [
-                [
-                    'pI',
-                ],
-            ],
-        ],
-        r: 'NU0',
+        H: 'I0',
+        r: 'H0',
     });
 });
 
@@ -185,14 +165,8 @@ test('Object-Wrapped Number: Encoding Expected -Infinity', (t) => {
     const source = new Number(-Infinity);
 
     t.deepEqual(testHelpers.simplifyEncoded(encode(source)), {
-        NU: [
-            [
-                [
-                    'nI',
-                ],
-            ],
-        ],
-        r: 'NU0',
+        H: 'J0',
+        r: 'H0',
     });
 });
 
@@ -202,14 +176,8 @@ test('Object-Wrapped Number: Encoding Expected NaN', (t) => {
     const source = new Number(NaN);
 
     t.deepEqual(testHelpers.simplifyEncoded(encode(source)), {
-        NU: [
-            [
-                [
-                    'Na',
-                ],
-            ],
-        ],
-        r: 'NU0',
+        H: 'C0',
+        r: 'H0',
     });
 });
 
@@ -219,13 +187,7 @@ test('Object-Wrapped Number: Encoding Expected -0', (t) => {
     const source = new Number(-0);
 
     t.deepEqual(testHelpers.simplifyEncoded(encode(source)), {
-        NU: [
-            [
-                [
-                    'n0',
-                ],
-            ],
-        ],
-        r: 'NU0',
+        H: 'M0',
+        r: 'H0',
     });
 });