+++ src/utils/genTypedArray.js
@@ -1,21 +1,21 @@
+import attachIndices from '/utils/attachIndices.js';
 import attachKeys from '/utils/attachKeys.js';
-import decodePointer from '/utils/decodePointer.js';
 import encodeWithAttachments from '/utils/encodeWithAttachments.js';
+import getSystemName from '/utils/getSystemName.js';
 
-export default (typeObj) => {
-    typeObj.D = {
-        _systemName: 'Date',
+export default (type) => {
+    return {
+        _systemName: getSystemName(new type()),
         _compressionType: 2,
         _encodeValue: (reference, attachments) => {
-            return encodeWithAttachments([[reference.valueOf()]], attachments);
+            return encodeWithAttachments([attachments._indices], attachments);
         },
         _generateReference: (store, dataItems) => {
-            return new Date(decodePointer(store, dataItems[0][0]));
+            return new type(dataItems[0].length);
         },
         _build: (store, dataItem) => {
+            attachIndices(store, dataItem);
             attachKeys(store, dataItem, 1, 2);
         },
     };
-
-    return typeObj;
 };