+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.?.?
+
+* Added BigInt64Array type support
+* Added BigUint64Array type support
+* Implemented Pointer-aware and data-aware encoding and decoding, drastically shrinking the size of the encoded output. See [Encoding](Encoding.md).
+
+
 ## 1.0.2 (December 11th, 2018)
 
 * Trying to get all NPM settings and tags correct